Split Worcester super-express and regular express
Even without the super-express it's still not possible to get a
half-hour repetition interval -- 40 minutes is the best we can do.
(This revision shows 45 minutes because that's better for customers.)
I suspect this is the source of some of the weird service patterns
Keolis operates, like how some of the "locals" still run express
through Newton.